Create utils/specifiers module and unit test detectLocalResolutionCollection
