+++ src/resolver.ts
@@ -7,23 +7,9 @@ import {
   serializeSpecifier
 } from '@glimmer/di';
 import { assert } from './utils/debug';
+import { detectLocalResolutionCollection } from './utils/specifiers';
 import { ModuleRegistry } from './module-registry';
 import { ResolverConfiguration } from './resolver-configuration';
-
-function detectLocalResolutionCollection(specifier: Specifier): string {
-  let { namespace, collection } = specifier;
-
-  // Look for the local-most private collection contained in the namespace
-  // (which will appear closest to the end of the string)
-  let startPos = namespace.lastIndexOf('/-');
-  if (startPos > -1) {
-    startPos += 2;
-    let endPos = namespace.indexOf('/', startPos);
-    collection = namespace.slice(startPos, endPos > -1 ? endPos : undefined);
-  }
-
-  return collection;
-}
 
 export default class Resolver implements IResolver {
   public config: ResolverConfiguration;
